Collapse My* task pages, share list-filter plumbing
MyTasksPage and MyCreatedTasksPage were ~250-line near-clones differing
only in view, storage keys, showAssignees, and i18n prefix — they are
now thin wrappers over one GlobalTasksPage taking those as concrete
props (routes untouched).

MyProjectsPage/MyDocumentsPage deliberately stay separate: beyond
columns and query hooks they diverge in guild-name resolution, debounce
handling, filter-bar i18n/ids, and fallbacks, so a shared page would
need a wide config surface. Only the byte-identical prefs/sort plumbing
(StoredPrefs, sanitizeStoredPrefs, useGlobalListFilters) is extracted.

Refs #869
